minor fixes and testing

+++ src/app/utils/directory.js
@@ -7,38 +7,26 @@ module.exports = {
             fs.mkdirSync(directoryPath);
         } catch (error) {
             if (error.code != 'EEXIST') {
-              console.error(`Error setting up the directory at ${directoryPath}:`, error);
+              console.error(`Error creating directory at ${directoryPath}:`, error);
               process.exit(1);
             }
         }
     },
 
     cleanDirectory(directoryPath) {
-        const files = fs.readdirSync(directoryPath);
-            for (let file of files) {
-                try {
-                    fs.unlinkSync(path.join(directoryPath, file));
+        if (fs.existsSync(directoryPath)) {
+            const files = fs.readdirSync(directoryPath);
+            if (files.length) {
+                try {    
+                    for (let file of files) {
+                        fs.unlinkSync(path.join(directoryPath, file));
+                    }
                 } catch (error) {
                     console.error(`Error removing the file ${file}:`, error);
                 }
+            } else {
+                console.log('Empty directory. Nothing to clean');
             }
-    },
-
-    existsDirectory(directoryPath) {
-        try {
-            fs.readdirSync(directoryPath);
-            return true;
-        } catch (error) {
-            return false;
-        }
-    },
-
-    setUpDirectory(directoryPath) {
-        try {
-            cleanDirectory(directoryPath);
-        } catch (error) {
-            console.log(`Directory ${directoryPath} not found. Creating directory ...`);
-            module.exports.createDirectory(directoryPath);
         }
     },
 }
